implements a general list view
instead of having different list views for each model, i created a list view, that renders the table spans depending on the values of the documents. For this to work, the documents are mapped before being passed to pug, because there are many internal properties on each document that i don't want to render.